+++ base/src/main/java/eu/fraho/spring/securityJwt/dto/RefreshToken.java
@@ -17,7 +17,7 @@ import lombok.Value;
 @EqualsAndHashCode(callSuper = true)
 public final class RefreshToken extends AbstractToken {
     @Builder
-    public RefreshToken(String token, int expiresIn) {
+    public RefreshToken(String token, long expiresIn) {
         super(token, expiresIn);
     }
